requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Database
psycopg2-binary>=2.9.9
supabase>=2.0.0

# Celery
celery>=5.3.4
redis>=5.0.0

# LangChain & AI
langchain>=0.1.0
langgraph>=0.0.20
langchain-google-genai>=0.0.6
langchain-openai>=0.0.5
langchain-anthropic>=0.0.1
langchain-community>=0.0.10

# Vector Store & Embeddings
pinecone-client>=3.0.0
sentence-transformers>=2.2.0

# Document Processing
pypdf2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
python-magic>=0.4.27

# Code Analysis & Processing
tree-sitter>=0.20.0
gitpython>=3.1.40
tiktoken>=0.5.2

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
requests>=2.31.0
aiohttp>=3.9.0

# Auth & Security
pyjwt>=2.8.0
bcrypt>=4.1.0
django-extensions>=3.2.3

# Testing
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
factory-boy>=3.3.0

# Code Quality
black>=23.12.0
flake8>=6.1.0
isort>=5.13.0

# Additional Django Tools
djangorestframework-simplejwt>=5.3.1
django-filter>=23.5
django-environ>=0.11.2

# WebSocket Support
websockets>=12.0

# Monitoring & Logging
sentry-sdk>=1.40.0
